Add files via upload
You have to address the strategies and data challenges you have faced in your project. The answers are fine, but you may like to brief why you had used this algorithm in the project or data analysis, along with its values in it. And also you have to explain more about univariate & bivariate analysis. Range estimate at 95%, confidence for the model performance in production, these are the missing details in your project. You need to provide the key hyperparameters available for the selected algorithm and regularization techniques used in your model. Also, you have to write the advantages & disadvantages of the algorithm used in your project, after that you have to compare your results with the base model. Keep up your good work